requirements.txt
Django==3.2.6
psycopg2-binary==2.9.1
drf-yasg==1.20.0
django-better-admin-arrayfield==1.4.2
python-decouple==3.4
django-timezone-field
django-mptt